+++ packages/snap-store-mobx/src/types.ts
@@ -63,33 +63,36 @@ export type FinderFieldConfig = {
 	levels?: string[];
 };
 
+export type AutocompleteStoreConfigSettings = {
+	integratedSpellCorrection?: boolean;
+	initializeFromUrl?: boolean;
+	syncInputs?: boolean;
+	serializeForm?: boolean;
+	facets?: FacetStoreConfig & {
+		fields?: {
+			[field: string]: FacetStoreConfig;
+		};
+	};
+	trending?: {
+		limit: number;
+		showResults?: boolean;
+	};
+	history?: {
+		limit: number;
+		showResults?: boolean;
+	};
+	redirects?: {
+		merchandising?: boolean;
+		singleResult?: boolean;
+	};
+};
+
 // Autocomplete config
 export type AutocompleteStoreConfig = StoreConfig & {
 	globals?: Partial<AutocompleteRequestModel>;
 	selector: string;
 	action?: string;
-	settings?: {
-		integratedSpellCorrection?: boolean;
-		initializeFromUrl?: boolean;
-		syncInputs?: boolean;
-		serializeForm?: boolean;
-		facets?: FacetStoreConfig & {
-			fields?: {
-				[field: string]: FacetStoreConfig;
-			};
-		};
-		trending?: {
-			limit: number;
-			showResults?: boolean;
-		};
-		history?: {
-			limit: number;
-			showResults?: boolean;
-		};
-		redirects?: {
-			merchandising?: boolean;
-		};
-	};
+	settings?: AutocompleteStoreConfigSettings;
 };
 
 // Recommendation config